backend/requirements.txt
flask>=2.3.0
flask-jwt-extended>=4.6.0
flask-cors>=4.0.0
aiohttp>=3.9.0
psycopg2-binary>=2.9.9
python-binance>=1.0.19
python-dotenv>=1.0.0
werkzeug>=3.0.0
gevent>=23.9.0
gunicorn>=21.2.0
pytrends>=4.9.2
